fix: address code review feedback — security, config, and schema alignment

+++ platform-web/backend/stock-api/src/main/java/com/stock/api/model/dto/MarketSummaryDTO.java
@@ -20,5 +20,5 @@ public class MarketSummaryDTO {
     @JsonProperty("total_volume")
     private Long totalVolume;
     @JsonProperty("total_amount")
-    private Long totalAmount;
+    private Double totalAmount;
 }
